Moving the scala sensor to its own package, and making the Measures completely functional
